Should we keep going?
I know that the game jam is over, but I'm curious to see what the
finished product would look like when it works. Do any of you want to
work with me to keep building on this and then share it to the club
sometime before the end of the semester?  If you do, just put another
commit in here after this one.  I'm just going to edit the size of one
of the images so that this can get through